% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAT_sampling_extend.R
\name{AAT_sampling_extend}
\alias{AAT_sampling_extend}
\title{Create an expanded sample \code{X_new} starting from a sample \code{X}}
\usage{
AAT_sampling_extend(X, distr_fun, distr_par, N_new, nrep = 10)
}
\arguments{
\item{X}{matrix (\code{N},\code{M}) of initial samples.}

\item{distr_fun}{string (eg: \code{"unif"}) if all inputs have the same pdf, or vector of \code{M} strings (eg: \code{c("unif", "norm")}) containing the probability distribution function of each input.}

\item{distr_par}{row vector if all input pdfs have the same parameters, list of \code{M} vectors otherwise containing the parameters of the probability distribution function.}

\item{N_new}{positive integer, new dimension of the sample (must be > \code{N})}

\item{nrep}{scalar, number of replicate to select the maximin hypercube (default value: 10)}
}
\value{
\code{X_new} matrix (\code{N_new}, \code{M}) of expanded sample.
}
\description{
This function create an expanded sample \code{X_new} starting from a sample \code{X} and using latin hypercube and the maximin criterion.
}
\examples{
# Example 1: 2 inputs, both from Unif[0,3]
N <- 1000 
M <- 2 
distr_fun <- "unif" 
distr_par <- c(0, 3)
samp_strat <- "lhs"
X <- AAT_sampling(samp_strat, M, distr_fun, distr_par, N)
# Plot results:
plot(X, xlab = expression(x[1]), ylab = expression(x[2]))
#
# Adding up new samples
N2 <- 500 # increase of base sample size
# (that means: N2 * (M+2) new samples that will need to be evaluated)
Xext <- AAT_sampling_extend(X, distr_fun, distr_par, 2 * (N + N2)) # extended sample 
# (it includes the already evaluated samples X and the new ones)
Xnew <- Xext[-(1:(2 * N)),] # extract the new input samples that need to be evaluated
# Plot results:
par(mfrow = c(1, 2))
plot(X, xlab = expression(x[1]), ylab = expression(x[2]), main = "X")
plot(Xnew, xlab = expression(x[1]), ylab = expression(x[2]), main = "X new")
}
\seealso{
\code{\link{AAT_sampling}}
}
