% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/above.R
\name{above}
\alias{above}
\title{This function can be used as input argument ("output_condition") when
applying pawn_indices, pawn_convergence, pawn_plot_ks}
\usage{
above(y, par)
}
\description{
Returns the indices of \code{y} above the scalar \code{par} threshold
}
