% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allrange.R
\name{allrange}
\alias{allrange}
\title{This function can be used as input argument ("output_condition") when
applying pawn_indices, pawn_convergence, pawn_plot_ks}
\usage{
allrange(y, par)
}
\description{
Returns all indices of \code{y}
This is often the default function
}
