% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/below.R
\name{below}
\alias{below}
\title{This function can be used as input argument ("output_condition") when
applying pawn_indices, pawn_convergence, pawn_plot_ks}
\usage{
below(y, par)
}
\description{
Returns the indices of \code{y} below the scalar \code{par} threshold
}
