% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot1_dummy.R
\name{boxplot1_dummy}
\alias{boxplot1_dummy}
\title{Boxplot of vectors
Boxplots when the mean, lower values and upper values are specified.}
\usage{
boxplot1_dummy(mu, lb = NULL, ub = NULL, prnam = NULL)
}
\arguments{
\item{mu}{vector (\code{M}) of mean or median values to be plotted}

\item{lb}{vector (\code{M}) of lower values to be plotted.}

\item{prnam}{labels for the x-axis of the boxplot}

\item{au}{vector (\code{M}) of upper values to be plotted.}
}
\description{
Boxplot of vectors
Boxplots when the mean, lower values and upper values are specified.
}
\examples{

# Setup the model and define input ranges
myfun  <- "ishigami_homma_function"
M <- 3
DistrFun <- "unif"
DistrPar <-  c(-pi, pi)
SampStrategy <- "lhs"
N <- 3000 # Base sample size.
X <- AAT_sampling(SampStrategy, M, DistrFun, DistrPar, N)
x_labels <- c('x(1)', 'x(2)', 'x(3)')

# Run the model 
Y <- model_evaluation(myfun, X) 

Compute indices with the PAWN method
n <- 10
pawn_ind <- pawn_indices(X, Y, n, dummy = TRUE)
KS_max <- pawn_ind$KS_max 
KS_dummy <- pawn_ind$KS_dummy 
KS_max_d <- c(KS_max,KS_dummy) 

# Plot results:
boxplot1_dummy(mu = KS_max_d, prnam = x_labels) + ylab("KS")
}
\seealso{
\code{\link{boxplot2}} \code{\link{boxplot}} \code{\link{plot}}
}
