% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot2.R
\name{boxplot2}
\alias{boxplot2}
\title{Boxplot of vectors with two groups
Boxplots when the mean, lower values and upper values are specified for two groups a and b.}
\usage{
boxplot2(
  a,
  b,
  al = NULL,
  au = NULL,
  bl = NULL,
  bu = NULL,
  labels = NULL,
  leg = NULL,
  col = 2:3,
  ylim = NULL,
  ...
)
}
\arguments{
\item{a}{vector (\code{M}) of mean values to be plotted}

\item{b}{vector (\code{M}) of mean values to be plotted}

\item{al}{vector (\code{M}) of lower values to be plotted. Default \code{al = NULL}}

\item{au}{vector (\code{M}) of upper values to be plotted. Default \code{au = NULL}}

\item{bl}{vector (\code{M}) of lower values to be plotted. Default \code{bl = NULL}}

\item{bu}{vector (\code{M}) of upper values to be plotted. Default \code{bu = NULL}}

\item{labels}{labels for the x-axis of the boxplot}

\item{leg}{names for the legend. Default \code{leg = NULL}}

\item{col}{colors for the plot. Default  \code{col = 2:3}}

\item{ylim}{limits of the y-axis. Default  \code{ylim = NULL}}

\item{...}{others arguments to be passed in \code{\link{plot}}}
}
\description{
Boxplot of vectors with two groups
Boxplots when the mean, lower values and upper values are specified for two groups a and b.
}
\examples{
# Setup the model and define input ranges
myfun  <- "ishigami_homma_function"
M <- 3
DistrFun <- "unif"
DistrPar <-  c(-pi, pi)
# Sample parameter space using the resampling strategy proposed by 
# (Saltelli, 2008; for reference and more details, see help of functions
# vbsa_resampling and vbsa_indices) 
SampStrategy <- "lhs"
N <- 3000 # Base sample size.
# Comment: the base sample size N is not the actual number of input 
# samples that will be evaluated. In fact, because of the resampling
# strategy, the total number of model evaluations to compute the two
# variance-based indices is equal to N*(M+2) 
X <- AAT_sampling(SampStrategy, M, DistrFun, DistrPar, 2 * N)
XABC <- vbsa_resampling(X)
# Run the model and compute selected model output at sampled parameter
# sets:
YA <- model_evaluation(myfun, XABC$XA) # size (N,1)
YB <- model_evaluation(myfun, XABC$XB) # size (N,1)
YC <- model_evaluation(myfun, XABC$XC) # size (N*M,1)
# Compute main (first-order) and total effects:
ind <- vbsa_indices(YA, YB, YC)
Si <- ind[1,]
STi <- ind[2,] 
# Plot results:
boxplot2(Si, STi, leg = c("main effects", "total effects"))
}
\seealso{
\code{\link{boxplot1}} \code{\link{boxplot}} \code{\link{plot}}
}
