% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EET_convergence.R
\name{EET_convergence}
\alias{EET_convergence}
\title{Mean and standard deviation of Elementary Effects}
\usage{
EET_convergence(EE, rr, Nboot = 0, alfa = 0.05)
}
\arguments{
\item{EE}{matrix (\code{r}, \code{M}) of \code{r} elementary effects}

\item{rr}{vector (\code{R}), sample sizes at which indices will be estimated (must be a vector of integer positive values not exceeding \code{r})}

\item{Nboot}{scalar, number of resamples used for boostrapping (default: 0)}

\item{alfa}{scalar, significance level for the confidence intervals estimated by bootstrapping (default: 0.05)}
}
\value{
List containing: 
\itemize{
  \item \code{m_r} mean of EEs at different sampling size
  \item \code{s_r} standard deviation of EEs at different sampling size
}
If \code{Nboot > 1} it also contains
\itemize{
  \item \code{m_lb_r} lower bound of EEs mean at different sampling size
  \item \code{m_ub_r} upper bound of EEs mean at different sampling size
  \item \code{s_lb_r} lower bound of EEs std.dev. at different sampling size
  \item \code{s_ub_r} upper bound of EEs std.dev. at different sampling size
  \item \code{m_sd_r} standard deviation of EEs mean at different sampling size
  \item \code{s_sd_r} standard deviation of EEs std.dev. at different sampling size. 
}
All output arguments are matrices of size (\code{R},\code{M})
}
\description{
Compute mean and standard deviation of Elementary Effects (EEs) using an increasing number of samples.
}
\examples{
# See the demo
# demo("workflow_eet_hymod")
# or
# demo("workflow_eet_hbv")
}
\seealso{
\code{\link{EET_indices}} \code{\link{EET_plot}}
}
