% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EET_indices.R
\name{EET_indices}
\alias{EET_indices}
\title{Mean and standard deviation of Elementary Effects}
\usage{
EET_indices(r, xrange, X, Y, design_type, Nboot = 0, alfa = 0.05)
}
\arguments{
\item{r}{scalar, number of sampling point}

\item{xrange}{list of \code{M} vectors of length 2 containing the input ranges}

\item{X}{matrix (\code{r * (M + 1)}, \code{M}) matrix of sampling datapoints where EE must be computed}

\item{Y}{vector (\code{r * (M + 1)}) vector of output values}

\item{design_type}{string, design type (string) Options: \code{"trajectory"}, \code{"radial"}}

\item{Nboot}{scalar, number of resamples used for boostrapping (default:0)    - scalar}

\item{alfa}{scalar, significance level for the confidence intervals estimated by bootstrapping (default: 0.05)}
}
\value{
List containing: 
\itemize{
  \item \code{mi} vector (\code{M}) mean of the elementary effects
  \item \code{sigma} vector (\code{M}) standard deviation of the elementary effects
  \item \code{EE} matrix (\code{r}, \code{M}) of elementary effects
}
If \code{Nboot > 1} it also contains
\itemize{
  \item \code{mi_sd} vector (\code{M}) standard deviation of \code{mi} across \code{Nboot} estimations
  \item \code{sigma_sd} vector (\code{M}) standard deviation of \code{sigma} across \code{Nboot} estimations
  \item \code{mi_lb} vector (\code{M}) lower bound of \code{mi} (at level \code{alfa}) across \code{Nboot} estimations
  \item \code{sigma_lb} vector (\code{M}) lower bound of \code{sigma} across \code{Nboot} estimations
  \item \code{mi_ub} vector (\code{M}) upper bound of \code{mi} (at level \code{alfa}) across \code{Nboot} estimations
  \item \code{sigma_ub} vector (\code{M}) upper bound of \code{sigma} across \code{Nboot} estimations. 
}
}
\description{
This function computes mean and standard deviation of Elementary Effects (EEs) \cr
Compute the sensitivity indices according to the Elementary Effects Test
(Saltelli, 2008) or 'method of Morris' (Morris, 1991). \cr
These are: the mean (\code{mi}) of the Elementary Effects (EEs) associated to
input 'i', which measures the input influence; and the standard deviation
(\code{sigma}) of the EEs, which measures its level of interactions with other inputs.
For the mean EE, we use the version suggested by Campolongo et al. (2007), 
where absolute values of the EEs are used (this is to avoid that EEs with opposite sign 
would cancel each other out).
}
\examples{
# See the demo
# demo("workflow_eet_hymod")
# or
# demo("workflow_eet_hbv")
}
\references{
Morris, M.D. (1991), Factorial sampling plans for preliminary computational experiments, Technometrics, 33(2). \cr
Saltelli, A., et al. (2008), Global Sensitivity Analysis, The Primer, Wiley. \cr
Campolongo, F., Cariboni, J., Saltelli, A. (2007), An effective screening design for sensitivity analysis of large models. Environ. Model. Softw. 22 (10), 1509-1518.
}
\seealso{
\code{\link{EET_convergence}} \code{\link{EET_plot}}
}
