% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EET_plot.R
\name{EET_plot}
\alias{EET_plot}
\title{EET plot
Plot the mean and the standard deviation of the elementary effects.}
\usage{
EET_plot(
  m,
  s,
  ml = NULL,
  mu = NULL,
  sl = NULL,
  su = NULL,
  labels = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = "Mean of EEs",
  ylab = "Sd of EEs",
  col = 1:length(m),
  pch = 15:19,
  ...
)
}
\arguments{
\item{m}{vector (\code{M}) mean of the elementary effects}

\item{s}{vector (\code{M}) standard deviation of the elementary effects}

\item{ml}{vector (\code{M}) lower bound of \code{mi} (at level \code{alfa}) across \code{Nboot} estimations. Default \code{ml = NULL}}

\item{mu}{vector (\code{M}) upper bound of \code{mi} (at level \code{alfa}) across \code{Nboot} estimations. Default \code{mu = NULL}}

\item{sl}{vector (\code{M}) lower bound of \code{sigma} across \code{Nboot} estimations. Default \code{sl = NULL}}

\item{su}{vector (\code{M}) upper bound of \code{sigma} across \code{Nboot} estimations. Default \code{su = NULL}}

\item{labels}{labels for legend. Default \code{labels = NULL}}

\item{xlim}{limits of the x-axis. Default \code{xlim = NULL}}

\item{ylim}{limits of the y-axis. Default  \code{ylim = NULL}}

\item{xlab}{label for the x-axis. Default \code{xlab = 'Mean of EEs'}}

\item{ylab}{label for the x-axis. Default \code{ylab = 'Sd of EEs'}}

\item{col}{colors for the plot. Default  \code{col = 1:length(m)}}

\item{pch}{symbols to use when plotting points. Default  \code{pch = 15:19}}

\item{...}{others arguments to be passed in \code{\link{plot}}}
}
\description{
EET plot
Plot the mean and the standard deviation of the elementary effects.
}
\examples{
# See the demo
# demo("workflow_eet_hymod")
# or
# demo("workflow_eet_hbv")
}
\seealso{
\code{\link{EET_indices}} \code{\link{EET_convergence}} \code{\link{plot}}
}
