% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAST_indices.R
\name{FAST_indices}
\alias{FAST_indices}
\title{Main effect (first-order) sensitivity index for the Fourier Amplitude Sensitivity Test (FAST)}
\usage{
FAST_indices(Y, M, Nharm = 4, omega = generate_FAST_frequency(M))
}
\arguments{
\item{Y}{vector (\code{N}), set of model output samples}

\item{M}{scalar, number of inputs}

\item{Nharm}{scalar, interference factor, i.e.the number of higher harmonics to be considered (default is 4)}

\item{omega}{vector (\code{M}), angular frequencies associated to inputs (default values computed by function \code{\link{generate_FAST_frequency}})}
}
\value{
List containing: 
\itemize{
  \item \code{Si} (vector of length \code{M}) of main effect (first-order) sensitivity indices
  \item \code{V} (scalar) total output variance
  \item \code{A} (vector \code{N}) Fourier coefficients
  \item \code{B} (vector \code{N}) Fourier coefficients
  \item \code{Vi} (vector of length \code{M}) output variances from each input 
}
}
\description{
Computes main effect (first-order) sensitivity index according to the Fourier Amplitude Sensitivity Test (FAST; Cukier et al., 1978).
}
\examples{
fun_test  <- 'sobol_g_function'
# Define input distribution and ranges:
M <- 5 # may range from 5 to 11
distr_fun <- 'unif'
distrpar <-  c(0, 1)
a <- 9

## Step 1: Choose sampling size # 

Nfast <- 2^5 + 1 # this value just for a quick example 
# it would be preferred: Nfast<- 2^13 + 1 
## Step 2: Approximate first-order variance-based indices by FAST

# FAST
Fsamp <- FAST_sampling(distr_fun, distrpar, M, N = Nfast)
X <- Fsamp$X
s <- Fsamp$s
# Run the model and compute selected model output at sampled parameter
# sets:
Y <- model_evaluation(fun_test, X, a = a)
# Estimate indices: 
Si_fast <- FAST_indices(Y, M)
}
\references{
Cukier, R.I., Levine, H.B., and Shuler, K.E. (1978), Nonlinear Sensitivity Analyis of Multiparameter Model SYstems, Journal of Computational Physics, 16, 1-42.

Saltelli, A., Tarantola, S. and Chan, K.P.S. (1999), A Quantitative Model-Independent Method ofr Global Sensitivty Analysis of Model Output, Technometrics, 41(1), 39-56.
}
\seealso{
\code{\link{FAST_sampling}} \code{\link{generate_FAST_frequency}}
}
