% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAST_sampling.R
\name{FAST_sampling}
\alias{FAST_sampling}
\title{Sampling for the Fourier Amplitude Sensitivity Test (FAST)}
\usage{
FAST_sampling(
  distr_fun,
  distr_par,
  M,
  Nharm = 4,
  omega = generate_FAST_frequency(M),
  N = 2 * Nharm * max(omega) + 1
)
}
\arguments{
\item{distr_fun}{string (eg: \code{'unif'}) if all inputs have the same pdf, or vector of \code{M} strings (eg: \code{c('unif','norm')}) containing the probability distribution function of each input.}

\item{distr_par}{row vector if all input pdfs have the same parameters, list of \code{M} vectors otherwise containing the parameters of the probability distribution function.}

\item{M}{scalar, number of inputs}

\item{Nharm}{scalar, interference factor, i.e.the number of higher harmonics to be considered (default is 4, taken from Saltelli et al. (1999; page 42))}

\item{omega}{vector (\code{M}), angular frequencies associated to inputs (default values computed by function \code{\link{generate_FAST_frequency}})}

\item{N}{odd scalar, number of samples (default is \code{2 * Nharm * max(omega) + 1} which is the minimum sampling size according to Cukier et al. (1978))}
}
\value{
\code{X} matrix of \code{N} input samples.
}
\description{
Implements sampling for the Fourier Amplitude Sensitivity Test (FAST; Cukier et al., 1978) and returns a matrix \code{X} of \code{N} input samples.
}
\examples{
fun_test  <- 'sobol_g_function'
# Define input distribution and ranges:
M <- 5 # may range from 5 to 11
distr_fun <- 'unif'
distrpar <-  c(0, 1)
a <- 9

## Step 1: Choose sampling size

# suggested values of Nfast

Nfast <- 2^13 + 1 
## Step 2: Approximate first-order variance-based indices by FAST

# FAST
Fsamp <- FAST_sampling(distr_fun, distrpar, M, N = Nfast)
X <- Fsamp$X
s <- Fsamp$s
}
\references{
Cukier, R.I., Levine, H.B., and Shuler, K.E. (1978), Nonlinear Sensitivity Analyis of Multiparameter Model SYstems, Journal of Computational Physics, 16, 1-42.

Saltelli, A., Tarantola, S. and Chan, K.P.S. (1999), A Quantitative Model-Independent Method ofr Global Sensitivty Analysis of Model Output, Technometrics, 41(1), 39-56.
}
\seealso{
\code{\link{FAST_sampling_unif}} \code{\link{generate_FAST_frequency}}
}
