% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_FAST_frequency.R
\name{generate_FAST_frequency}
\alias{generate_FAST_frequency}
\title{Generate FAST frequency}
\usage{
generate_FAST_frequency(M)
}
\arguments{
\item{M}{integer scalar between 4 and 50, number of inputs}
}
\description{
Generate frequency set free of interferences through (at least) 4th order (vector (\code{M}))
}
\details{
For \code{M > 4}, frequencies are computed based on the recursive algorithm by Cukier et al. (1975) which is free of interferences through the 4th order. 
For \code{M <= 4}, we use values from the literature that guarantee higher order interferences free: if \code{M = 2} use values from Sec. 3.1 in Xu, C. and G. Gertner (2007) (free of interference through 10th order), if \code{M = 4} use values from Table III in Cukier et al. (1975) (free of interferences through 6th order).
}
\examples{
M <- 5
generate_FAST_frequency(M)
}
\references{
Cukier et al. (1975) Study of the sensitivity of coupled reaction systems to uncertainties in rate coefficients. III. Analysis of the  approximations, J. Chem. Phys. 63, 1140

Xu, C. and G. Gertner (2007), Extending a global sensitivity analysis technique to models with correlated parameters, Computational Statistics and Data Analysis, 51, 5579-5590.
}
