% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbv_snow_objfun.R
\name{hbv_snow_objfun}
\alias{hbv_snow_objfun}
\title{hbv snow objfun}
\usage{
hbv_snow_objfun(param, dat, warmup, Case)
}
\arguments{
\item{param}{vector (13) of model parameters. 
 Snow routine parameters:
\itemize{
  \item \code{Ts} = threshold temperature [C]
  \item \code{CFMAX} = degree day factor [mm/C]
  \item \code{CFR} = refreezing factor [-]
  \item \code{CWH} = Water holding capacity of snow [-]
  \item \code{temp} = temperature [C]
} 
HBV parameters: 
\itemize{
  \item \code{BETA} = Exponential parameter in soil routine [-]
  \item \code{LP} = evapotranspiration limit [-]
  \item \code{FC} = field capacity [mm]
  \item \code{PERC} = maximum flux from Upper to Lower Zone [mm/Dt]
  \item \code{K0} = Near surface flow coefficient (ratio) [1/Dt]
  \item \code{K1} = Upper Zone outflow coefficient (ratio) [1/Dt]
  \item \code{K2} = Lower Zone outflow coefficient (ratio) [1/Dt]
  \item \code{UZL} = Near surface flow threshold [mm]
}}

\item{dat}{dataset containing time series of precipitation (\code{prec}), evapotranspiration (\code{ept}), observed flow (\code{flow}) and temperature (\code{temp})}

\item{warmup}{scalar, warmup time}

\item{Case}{scalar, 1 or 2, indicates the preferred path. \code{Case = 1}: runoff from the upper zone, \code{Case = 2}: percolation.}
}
\value{
List containing: 
\itemize{
  \item \code{f} = vector (6) of objective functions 1: AME, 2: NSE, 3: BIAS, 4: TRMSE, 5: SFDCE, 6: RMSE
  \item  \code{Q_sim} = vector (\code{N}) time series of simulated flow
  \item \code{STATES} = matrix \code{(N, 5)} time series of simulated storages (all in mm) 1: water content of snowpack (snow component) 2: water content of snowpack (liquid component) 3: water content of soil (soil moisture) 4: water content of upper reservoir of flow routing routine 5: water content of lower reservoir of flow routing routine
  \item \code{FLUXES} = matrix \code{(N, 8)} time series of simulated fluxes (all in mm/Dt) 1: refreezing 2: snowmelt 3: actual evapotranspiration 4: recharge (water flux from soil moisture accounting module to flow routing module) 5: percolation (water flux from upper to lower reservoir of the flow routing module) 6: runoff from upper reservoir 7: runoff from lower reservoir
}
}
\description{
This function simulates the snow accumulation/melting process (via the internal function \code{snow_routine} and the rainfall-runoff process (via the HBV model by Seibert (1997)) and returns 6 objective functions (see Kollat et al, 2002).
}
\references{
Seibert,J.(1997)."Estimation of Parameter Uncertainty in the HBV Model". Nordic Hydrology.28(4/5).247-262.

Kollat,J.B.,Reed,P.M.,Wagener,T.(2002)."When are multiobjective calibration trade-offs in hydrologic models meaningful?". Water resources research, VOL.48, W03520.
}
