% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbvdata.R
\docType{data}
\name{hbvdata}
\alias{hbvdata}
\title{hbvdata}
\format{
Data frame with 20454 rows and 7 columns
}
\description{
The case study area is is the Nezinscot River at Turner center, Maine, USA (USGS 01055500, see \url{http://waterdata.usgs.gov/nwis/nwismap})
}
\details{
\itemize{
  \item \code{date} vector \code{T} of dates of the time series
  \item \code{prec} vector \code{T} time series of precipitation
  \item \code{ept} vector \code{T} time series of evapotranspiration
  \item \code{flow} vector \code{T} time series of observed flow 
  \item \code{temp} vector \code{T} time series of temperature, it is the mean of \code{t1} and \code{t2}
  \item \code{t1} vector \code{T} time series of temperature \code{t1}
  \item \code{t2} vector \code{T} time series of temperature \code{t2}
}
}
\keyword{datasets}
