% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hymod_MulOut.R
\name{hymod_MulOut}
\alias{hymod_MulOut}
\title{Hymod RMSE, BIAS, MEAN, STANDARD DEVIATION AND VARIANCE of the simulated flow}
\usage{
hymod_MulOut(x, dat)
}
\arguments{
\item{x}{vector \code{5} of model parameters \code{(Smax, beta, alfa, Rs, Rf)}}

\item{dat}{dataset containing \code{rain} vector \code{T} time series of rainfall, \code{evap} vector \code{T} time series of potential evaporation, \code{flow} vector \code{T} time series of observed flow.}
}
\value{
\code{Y}  vector \code{(5)} of objective functions \code{(RMSE, BIAS, MEAN(Qs), STD(Qs), VAR(Qs))}
}
\description{
This function runs the rainfall-runoff Hymod model and returns 2 metrics of model performance: RMSE and BIAS
}
\seealso{
\code{\link{hymod_sim}}
}
