% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hymod_sim.R
\name{hymod_sim}
\alias{hymod_sim}
\title{Hymod rainfall-runoff model}
\usage{
hymod_sim(rain, evap, param)
}
\arguments{
\item{rain}{vector \code{(T)} time series of rainfall}

\item{evap}{vector \code{(T)} time series of potential evaporation}

\item{param}{5 elements vector of model parameters \code{(Smax, beta, alfa, Rs, Rf)}}
}
\description{
This function simulates the Hymod rainfall-runoff model
[Boyle, 2001; Wagener et al., 2002]
}
\references{
Boyle, D. (2001). Multicriteria calibration of hydrological models. PhD thesis, Dep. of Hydrol. and Water Resour., Univ. of Ariz., Tucson.

Wagener, T., Boyle, D., Lees, M., Wheater, H., Gupta, H., and Sorooshian, S. (2001). A framework for development and application of hydrological models. Hydrol. Earth Syst. Sci., 5, 13-26.
}
