% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ishigami_homma_noise_function.R
\name{ishigami_homma_noise_function}
\alias{ishigami_homma_noise_function}
\title{Ishigami Homma function}
\usage{
ishigami_homma_noise_function(x)
}
\arguments{
\item{x}{vector \code{(M)} of inputs}
}
\value{
\code{Y} matrix \code{(N, P)} of associated model ouputs (\code{P} being the number of scalar model outputs associated to each sampled input combination), \code{comp_time} total computing time for model evaluation (sec)
}
\description{
This function does  Ishigami Homma
}
\examples{
# See the demo
# demo("workflow_visual_ishigami_homma")
# or 
# demo("workflow_vbsa_ishigami_homma")
}
