% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeafCatch.R
\docType{data}
\name{LeafCatch}
\alias{LeafCatch}
\title{Leaf Catch}
\format{
Data frame with 14610 rows and 3 columns
}
\description{
Leaf Catch daily rainfall, evaporation and flow.
}
\details{
\itemize{
  \item rain vector \code{(T)} time series of rainfall (mm/h). 	
  \item evaporation vector \code{(T)} time series of potential evaporation (mm/h). 
  \item flow vector \code{T} time series of observed flow (m3/s).
}
}
\examples{
data(LeafCatch)
}
\keyword{datasets}
