% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Morris_orientation_matrix.R
\name{Morris_orientation_matrix}
\alias{Morris_orientation_matrix}
\title{Morris orientation matrix}
\usage{
Morris_orientation_matrix(k, p)
}
\arguments{
\item{k}{integer positive scalar, number of inputs}

\item{p}{integer positive even scalar, number of levels}
}
\value{
\code{Bstar} matrix of \code{(k + 1)} datapoints in the \code{k}-dimensional input space (to be used for computing one Elementary Effect for each of the \code{k} inputs)
}
\description{
This function builds a Morris orientation matrix
}
\examples{
# Example in two-dimensional space (k = 2):
#
p <- 4
plot(1, xlim = c(0,1), ylim = c(0, 1), type ="n", xlab = "", ylab = "")
Bstar <- Morris_orientation_matrix(2,p)
lines(Bstar[,1], Bstar[,2], col = "red")
points(Bstar[,1],Bstar[,2], col = "red", pch = 20)
# if you want to generate more datapoints:
Bstar <- Morris_orientation_matrix(2,p)
lines(Bstar[,1], Bstar[,2], col = "red")
points(Bstar[,1],Bstar[,2], col = "red", pch = 20)
Bstar <- Morris_orientation_matrix(2,p)
lines(Bstar[,1], Bstar[,2], col = "red")
points(Bstar[,1],Bstar[,2], col = "red", pch = 20)
Bstar <- Morris_orientation_matrix(2,p)
lines(Bstar[,1], Bstar[,2], col = "red")
points(Bstar[,1],Bstar[,2], col = "red", pch = 20)
}
