% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Morris_sampling.R
\name{Morris_sampling}
\alias{Morris_sampling}
\title{Morris Sampling}
\usage{
Morris_sampling(r, xmin, xmax, L)
}
\arguments{
\item{r}{positive integer number, number of elementary effects.}

\item{xmin}{vector \code{(M)}. Lower bounds of input ranges.}

\item{xmax}{vector \code{(M)}. Upper bounds of input ranges.}

\item{L}{positive, even number. Number of levels in the sampling grid}
}
\value{
\code{X} matrix of sampling datapoints where \code{EE} must be computed. This is a matrix with \code{r * (M + 1)} rows and \code{M} columns. Each row is a point in the input space. Rows are sorted in \code{r} blocks, each including \code{M + 1} rows. Within each block, points (rows) differ in one component at the time. Thus, each block can be used to compute one Elementary Effect \code{(EE_i)} for each model input \code{(i = 1, ...,M)}.
}
\description{
This function builds a matrix X of input samples to be used for the Elementary Effects Test, using a One-At-the-Time sampling strategy as described in Campolongo et al. (2011).
}
\references{
Morris, M.D. (1991), Factorial sampling plans for preliminary computational experiments, Technometrics, 33(2).
}
