% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawn_CDF.R
\name{pawn_CDF}
\alias{pawn_CDF}
\title{Compute the CDFs for the PAWN method}
\usage{
pawn_CDF(X, Y, n = 10, dummy = FALSE, verbose = TRUE)
}
\arguments{
\item{X}{matrix \code{(N, M)} set of inputs samples}

\item{Y}{matrix \code{(N, 1)} set of output samples}

\item{n}{number of conditional intervals (default: 10) \cr
- integer if all inputs have the same number of groups \cr
- list of \code{M} integers otherwise}

\item{dummy}{boolean, if dummy is True, an articial input is added to the 
set of inputs and the sensitivity indices are calculated for the dummy input. 
See the help of \link{pawn_indices} for refence and further explanation 
on the usage of the dummy input.}

\item{verbose}{boolean, provides additional information on how the samples are split by the function \link{pawn_split_sample}}
}
\value{
List containing: 
\itemize{
  \item \code{YF} vector \code{(P, 1)}, values of Y at which the Cumulative  Distribution Functions (CDFs) FU and FC are given 
  \item \code{FU} list of \code{M} elements, values of the empirical unconditional output CDFs.
  \item \code{FC} list of lists, values of the empirical conditional output CDFs for each input and each conditioning interval.
  \code{FC[[i]]} is a list of \code{n_eff[i]} CDFs conditional to the i-th input.
  \code{FC[[i]][k]} is obtained by fixing the i-th input to its k-th conditioning interval (while the other inputs vary freely)
  \item \code{xc}  list of \code{M} elements, subsamples' centers (i.e. mean value of the \code{Xi} over each conditioning interval)
  \code{xc[[i]]} vector of length \code{n_eff[i]} contains the centers of the \code{n_eff[i]} conditioning intervals for the i-th input.
  \item \code{idx_bootsize_min} scalar, used to estimate the smaller sample size across inputs, 
  so that the sensitivity index for the dummy input estimates the 'worst' approximation error 
  of the sensitivity index across the inputs.
  \item \code{FC_dummy} vector \code{(P, 1)}, conditional CDF for the dummy input.
}
}
\description{
This function computes the unconditional output Cumulative Distribution Funtions 
(i.e. when all inputs vary) and the conditional CDFs (when one input is fixed to a given 
conditioning interval, while the other inputs vary freely).
}
\details{
The function splits the output sample to create the conditional output by calling internally 
the function \link{pawn_split_sample}. 
The splitting strategy is an extension of the strategy for uniformy distributed inputs 
described in Pianosi and Wagener (2018) to handle inputs sampled from any distribution.
(see help of \link{pawn_split_sample} for further explanation).

The sensitivity indices for the PAWN method (KS statistic) measures the distance between 
these conditional and unconditional output CDFs 
(see help of \link{pawn_indices} for further details and reference).
}
\references{
Pianosi, F. and Wagener, T. (2018), Distribution-based sensitivity analysis from a generic input-output sample, Env. Mod. & Soft., 108, 197-207. \cr
}
