% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawn_convergence.R
\name{pawn_convergence}
\alias{pawn_convergence}
\title{This function computes the PAWN sensitivity indices  
using sub-samples of the original sample \code{Y} of increasing size.}
\usage{
pawn_convergence(
  X,
  Y,
  n = 10,
  NN,
  Nboot = 1,
  dummy = FALSE,
  output_condition = allrange,
  par = list()
)
}
\arguments{
\item{X}{matrix \code{(N, M)} set of inputs samples}

\item{Y}{matrix \code{(N, 1)} set of output samples}

\item{n}{number of conditional intervals (default: 10) \cr
- integer if all inputs have the same number of groups \cr
- list of \code{M} integers otherwise}

\item{NN}{vector \code{(R)} of subsample sizes at which indices will be estimated (\code{max(NN)} must not exceed \code{N})}

\item{Nboot}{scalar, number of bootstrap resamples to derive confidence intervals}

\item{dummy}{boolean, if \code{dummy} is TRUE, an artificial input is added to the set of inputs 
and the sensitivity indices are calculated for the dummy input.
See help of \link{pawn_indices} for reference and further explanation on the usage of the dummy input. \cr
ADVANCED USAGE: \cr
for Regional-Response Global Sensitivity Analysis: \cr}

\item{output_condition}{function (see the help for \link{pawn_indices} for information on this optional input)}

\item{par}{list (see the help for \link{pawn_indices} for information on this optional input)}
}
\value{
List containing:
\itemize{
  \item \code{KS_median} list of shape \code{(Nboot, M)} median KS across the conditioning intervals 
  (one value for each input and each bootstrap resample)
  \item \code{KS_mean} list of shape \code{(Nboot, M)} mean KS across the conditioning intervals 
  (one value for each input and each bootstrap resample)
  \item \code{KS_max} list of shape \code{(Nboot, M)} max KS across the conditioning intervals 
  (one value for each input and each bootstrap resample)
  \item \code{KS_dummy} list of shape \code{(Nboot, 1)} KS of dummy input (one value for each bootstrap resample)
  }
  @seealso \code{\link{pawn_split_sample}} \code{\link{pawn_indices}}
}
\description{
The function splits the output sample to create the conditional output.
The splitting strategy is an extension of the strategy for uniformy
distributed inputs described in Pianosi and Wagener (2018) to handle inputs sampled from 
any distribution.
}
\references{
Pianosi, F. and Wagener, T. (2018), Distribution-based sensitivity analysis from a generic input-output sample, Env. Mod. & Soft., 108, 197-207. \cr
Pianosi, F. and Wagener, T. (2015), A simple and efficient method for global sensitivity analysis based on cumulative distribution functions, Env. Mod. & Soft., 67, 1-11. \cr
Khorashadi Zadeh et al. (2017), Comparison of variance-based and moment-independent global sensitivity analysis approaches by application to the SWAT model, Environmental Modelling & Software,91, 210-222.
}
