% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawn_ks.R
\name{pawn_ks}
\alias{pawn_ks}
\title{Compute the Kolmogorov-Smirnov (KS)}
\usage{
pawn_ks(YF, FU, FC, output_condition = allrange, par = list())
}
\arguments{
\item{YF}{vector \code{(P, 1)}, values of Y at which the Cumulative  Distribution Functions (CDFs) FU and FC are given}

\item{FU}{list of \code{M} elements, values of the empirical unconditional output CDFs.}

\item{FC}{list of lists, values of the empirical conditional output CDFs for each input and each conditioning interval.
\code{FC[[i]]} is a list of \code{n_eff[i]} CDFs conditional to the i-th input.
\code{FC[[i]][k]} is obtained by fixing the i-th input to its k-th conditioning interval (while the other inputs vary freely)

ADVANCED USAGE: \cr
for Regional-Response Global Sensitivity Analysis: \cr}

\item{output_condition}{function (see the help for \link{pawn_indices} for information on this optional input)}

\item{par}{list (see the help for \link{pawn_indices} for information on this optional input)}
}
\value{
\code{KS} list of \code{M} elements. KS-statistic calculated between conditional and unconditional output for the M inputs 
and the \code{n_eff} conditioning intervals. \code{KS[i]} contains the KS values for the i-th input and the 
\code{n_eff[i]} conditioning intervals.
}
\description{
This function computes the KS between the unconditional and conditional output 
for each input and each conditioning interval, for ONE sample/bootstrap resample. \cr
This function is called internally in \link{pawn_indices} and \link{pawn_plot_ks}. \cr
The inputs necessary to call this function are estimated by \link{pawn_CDF}.
}
