% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawn_plot_CDF.R
\name{pawn_plot_CDF}
\alias{pawn_plot_CDF}
\title{Plot the CDFs estimated by the PAWN method}
\usage{
pawn_plot_CDF(
  X,
  Y,
  n = 10,
  n_col = 5,
  y_label = "output y",
  labelinput = "",
  verbose = TRUE
)
}
\arguments{
\item{X}{matrix \code{(N, M)} set of inputs samples}

\item{Y}{matrix \code{(N, 1)} set of output samples}

\item{n}{number of conditional intervals (default: 10) \cr
- integer if all inputs have the same number of groups \cr
- list of \code{M} integers otherwise}

\item{n_col}{integer, number of panels per row in the plot (default: \code{min(5, M))}}

\item{y_label}{string, legend for the horizontal axix (default: 'output y')}

\item{labelinput}{vector, label the vertical axis (default: \code{["X1","X2",...,"XM"]})}

\item{verbose}{boolean, provides additional information on how the samples are split by the function \link{pawn_split_sample}}
}
\value{
List containing: 
\itemize{
  \item \code{YF} vector \code{(P, 1)}, values of Y at which the Cumulative  Distribution Functions (CDFs) FU and FC are given 
  \item \code{FU} list of \code{M} elements, values of the empirical unconditional output CDFs.
  \item \code{FC} list of lists, values of the empirical conditional output CDFs for each input and each conditioning interval.
  \code{FC[[i]]} is a list of \code{n_eff[i]} CDFs conditional to the i-th input.
  \code{FC[[i]][k]} is obtained by fixing the i-th input to its k-th conditioning interval (while the other inputs vary freely)
  \item \code{xc}  list of \code{M} elements, subsamples' centers (i.e. mean value of the \code{Xi} over each conditioning interval)
\code{xc[[i]]} vector of length \code{n_eff[i]} contains the centers of the \code{n_eff[i]} conditioning intervals for the i-th input.
}
}
\description{
This function plots the unconditional output Cumulative Distribution Funtions 
(i.e. when all inputs vary) and the conditional CDFs (when one input is fixed 
to a given conditioning interval, while the other inputs vary freely).
}
\details{
The unconditional and conditional CDF are computed by the function \link{pawn_CDF} \cr
See the help of \link{pawn_CDF} for further explanation.
}
