% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawn_split_sample.R
\name{pawn_split_sample}
\alias{pawn_split_sample}
\title{PAWN split sample}
\usage{
pawn_split_sample(X, Y, n = 10, verbose = TRUE)
}
\arguments{
\item{X}{matrix \code{(N, M)} set of inputs samples}

\item{Y}{matrix \code{(N, 1)} set of output samples}

\item{n}{number of conditional intervals (default: 10) \cr
- integer if all inputs have the same number of groups \cr
- list of \code{M} integers otherwise}

\item{verbose}{boolean, provides additional information on how the samples are split}
}
\value{
List containing: 
\itemize{
  \item \code{YY} list of \code{M} elements. Output samples to assess the conditional CDFs. 
  \code{YY[[i]]} is a list of \code{n_eff[i]} subsamples that can be used to assess \code{n_eff[i]}
  conditional distributions with respect to the i-th input.
  \code{YY[[i]][k]} is obtained by fixing the i-th input to its k-th conditioning interval 
  (while the other inputs vary freely), and its length is \code{NC[[i]][k]}
  \item \code{xc} list of \code{M} elements. Subsamples' centers (i.e. mean value of \code{Xi} over each conditioning interval).
  The vector \code{xc[[i]]} of dimensions \code{(n_eff[i],1)} contains the subsamples' centers of the i-th input \code{Xi}.
  \item \code{NC} list of \code{M} elements. Number of data points in each conditioning interval and for each input.
  \code{NC[[i]]} has length \code{n_eff[i]} and contains the sample sizes for the i-th input.
  \item \code{n_eff} list of \code{M} elements. Number of conditioning intervals actually used 
  for each inputs \code{Xi} (see (*) for further explanation).
  \item \code{Xk} list of \code{M} elements. Subsamples' edges for each input \code{Xi} 
  (i.e. bounds of \code{Xi} over each conditioning interval)
  \code{Xk[i]} is a list of length \code{n_eff[i]+1} and contains the edges of the conditioning intervals 
  for the i-th input.
  \item \code{XX} list of \code{M} elements. Conditional input samples corresponding to the samples in \code{YY}.
  \code{XX[[i]]} is a list of \code{n_eff[i]} subsamples for the i-th input.
  \code{XX[[i]][k]} is obtained by fixing the i-th input to its k-th conditioning interval (while the other inputs vary freely), 
  and it has shape \code{(NC[[i]][k],M])}
  } \cr
  NOTES: \cr
(*)  - When \code{Xi} is discrete and when the number of values taken by \code{Xi} \code{(nxi)} is
lower than the prescribed number of conditioning intervals \code{(n[i])}, 
a conditioning interval is created for each value of \code{Xi} (and therefore the actual number 
of conditioning intervals is set to \code{n_eff[i] = nxi}). \cr
- The function ensures that values of \code{Xi} that are repeated several times
belong to the same group. This may lead to a number of conditioning
intervals \code{n_eff[i]} lower than the prescribed value \code{n[i]} and to a 
different number of data points between the groups.
}
\description{
This function splits a generic input-output dataset in order to create unconditional and conditional samples for the approximation of PAWN sensitivity indices (Pianosi and Wagener, 2018) \cr
This function extends the splitting strategy described in (Pianosi and Wagener, 2018), which consists of 
splitting the input and output sample into a number of equally spaced conditioning intervals (intervals with the same size) 
based on the value of each input \code{Xi}. Here, the conditioning intervals are equiprobable, i.e. there is
(approximately) the same number of data points in each interval. While equally spaced intervals are adapted for 
inputs that have a uniform distribution, equiprobable intervals allow to handle inputs that have any distribution. \cr
This function is called internally in \link{pawn_indices} and \link{pawn_plot_CDF}
}
\references{
Pianosi, F. and Wagener, T. (2018), Distribution-based sensitivity analysis from a generic input-output sample, Env. Mod. & Soft.
}
