% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_convergence_thres.R
\name{RSA_convergence_thres}
\alias{RSA_convergence_thres}
\title{Regional Sensitivity Analysis (with threshold)}
\usage{
RSA_convergence_thres(
  X,
  Y,
  NN,
  threshold = NULL,
  flag = 1,
  Nboot = 0,
  alfa = 0.05
)
}
\arguments{
\item{X}{matrix \code{(N, M)} set of inputs samples}

\item{Y}{matrix \code{(N, P)} set of output samples}

\item{NN}{vector \code{(R)} of subsample sizes at which indices will be estimated (\code{max(NN)} must not exceed \code{N})}

\item{threshold}{vector \code{(P)} threshold for output values. Default value: \code{threshold = median(Y)}}

\item{flag}{scalar specify the statistic to assess the distance between CDFs                                   \code{flag = 1}: \code{stat} maximum vertical distance (see \code{mvd} below), \code{flag = 2}: \code{stat} area between curves (see \code{spread} below) \code{flag = 3}: \code{stat} input range reduction (see \code{irr} below) Default value: \code{flag = 1}}

\item{Nboot}{scalar, number of resamples used for boostrapping. Default \code{Nboot = 0}, i.e. no bootstrapping.}

\item{alfa}{scalar significance level for the confidence intervals estimated by bootstrapping. Default: \code{0.05}}
}
\value{
List containing: 
\itemize{
  \item \code{stat} distance measure between CDFs for each input vector \code{(M)}. 
}
If \code{Nboot > 1} it also contains
\itemize{ 
  \item \code{stat_lb} lower bound of vector \code{(stat)} from bootstrapping vector \code{(M)}
  \item \code{stat_ub} upper bound of \code{stat} from bootstrapping vector \code{(M)}
}
}
\description{
This function computes and plots the sensitivity indices obtained by \code{\link{RSA_indices_thres}} using an increasing number of output samples.
}
\examples{
# See the demo
# demo("workflow_rsa_hymod")
}
\seealso{
\code{\link{RSA_plot_thres}} \code{\link{RSA_indices_thres}}
}
