% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_indices_groups.R
\name{RSA_indices_groups}
\alias{RSA_indices_groups}
\title{Regional Sensitivity Analysis with grouping}
\usage{
RSA_indices_groups(X, Y, ngroup = 10, Nboot = 0)
}
\arguments{
\item{X}{matrix \code{(N, M)} set of inputs samples}

\item{Y}{matrix \code{(N, P)} set of output samples}

\item{ngroup}{integer, number of groups considered (default: 10)}

\item{Nboot}{scalar, number of resamples used for boostrapping. Default \code{Nboot = 0}, i.e. no bootstrapping.}
}
\value{
List containing: 
\itemize{
\item \code{mvd_median} vector \code{(M)} if Nboot <= 1 or matrix \code{(Nboot,M)} if Nboot > 1, median of mvd between pairs of inputs' CDFs estimated for 
the different sub-sets and for each bootstrap resample when bootstrapping is used (i.e.Nboot > 1)
\item \code{mvd_mean} vector \code{(M)} if Nboot <= 1 or matrix \code{(Nboot,M)} if Nboot > 1, mean of mvd between pairs of inputs' CDFs estimated for 
the different sub-sets and for each bootstrap resample when bootstrapping is used (i.e.Nboot > 1)
\item \code{mvd_max} vector \code{(M)} if Nboot <= 1 or matrix \code{(Nboot,M)} if Nboot > 1, max of mvd between pairs of inputs' CDFs estimated for 
the different sub-sets and for each bootstrap resample when bootstrapping is used (i.e.Nboot > 1)
\item \code{spread_median} vector \code{(M)} if Nboot <= 1 or matrix \code{(Nboot,M)} if Nboot > 1, median of spread between pairs of inputs' CDFs estimated for 
the different sub-sets and for each bootstrap resample when bootstrapping is used (i.e.Nboot > 1)
\item \code{spread_mean} vector \code{(M)} if Nboot <= 1 or matrix \code{(Nboot,M)} if Nboot > 1, mean of spread between pairs of inputs' CDFs estimated for 
the different sub-sets and for each bootstrap resample when bootstrapping is used (i.e.Nboot > 1)
\item \code{spread_max} vector \code{(M)} if Nboot <= 1 or matrix \code{(Nboot,M)} if Nboot > 1, max of spread between pairs of inputs' CDFs estimated for 
the different sub-sets and for each bootstrap resample when bootstrapping is used (i.e.Nboot > 1)
\item \code{idxb} vector \code{N} respective group of the samples
\item \code{Yk} vector \code{ngroup + 1} range of \code{Y} in each group
}
You can easily derive the n_groups datasets \code{Xi} as: \code{Xi = X[idx == i]}

\cr
NOTES: \cr
- When \code{Y} is discrete and when the number of values taken by \code{Y} (ny) is lower than 
the prescribed number of groups (\code{ngroup}), a group is created for each value of \code{Y} 
(and therefore the number of groups is set to ny). \cr
The function ensures that values of Y that are repeated several times belong to the same group. 
This may lead to a final number of group lower than \code{ngroup} and to a different number of 
data points across the groups.
}
\description{
Computation function for Regional Sensitivity Analysis with grouping (as first proposed by Wagener et al. 2001). \cr
The function can handle discrete outputs. \cr
The function splits the samples in a dataset X into \code{ngroup} sub-sets corresponding to \code{ngroup} of equal size 
based on the value of \code{Y} (i.e. 'equiprobable' groups). \cr
Then it assesses the distance (i.e. maximum vertical distance called 'mvd' and area between CDFs called 'spread') 
between pairs of CDFs of \code{X} in the different sub-sets. It aggregates the values using a statistic (median,
mean and maximum) e.g. for mvd the function computes:
}
\details{
\eqn{mvd_median = median( max_x( | Fi(x) - Fj(x) | ) )}

\eqn{mvd_mean = mean( max_x( | Fi(x) - Fj(x) | ) )}

or

\eqn{mvd_max = max( max_x( | Fi(x) - Fj(x) | ) )}

where \eqn{Fi()} is the CDF of \code{X} in i-th dataset and \eqn{Fj()} is the CDF in the
j-th group.
}
\examples{
# See the demo
# demo("workflow_rsa_hymod")
}
\references{
Wagener, T., Boyle, D. P., Lees, M. J., Wheater, H. S., Gupta, H. V., and Sorooshian, S. (2001): A framework for development and application of hydrological models, Hydrol. Earth Syst. Sci., 5, 13-26.
}
\seealso{
\code{\link{RSA_plot_groups}} \code{\link{RSA_indices_thres}}
}
