% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_indices_thres.R
\name{RSA_indices_thres}
\alias{RSA_indices_thres}
\title{Regional Sensitivity Analysis (with threshold)}
\usage{
RSA_indices_thres(X, Y, threshold = NULL, flag = 1, Nboot = 0, alfa = 0.05)
}
\arguments{
\item{X}{matrix \code{(N, M)} set of inputs samples}

\item{Y}{matrix \code{(N, P)} set of output samples}

\item{threshold}{vector \code{(P)} threshold for output values. Default value: \code{threshold = median(Y)}}

\item{flag}{scalar specify the statistic to assess the distance between CDFs                                   \code{flag = 1}: \code{stat} maximum vertical distance (see \code{mvd} below), \code{flag = 2}: \code{stat} area between curves (see \code{spread} below) \code{flag = 3}: \code{stat} input range reduction (see \code{irr} below) Default value: \code{flag = 1}}

\item{Nboot}{scalar, number of resamples used for boostrapping. Default \code{Nboot = 0}, i.e. no bootstrapping.}

\item{alfa}{scalar significance level for the confidence intervals estimated by bootstrapping. Default: \code{0.05}}
}
\value{
List containing: 
\itemize{
  \item \code{stat} vector \code{(M)} distance measure between CDFs for each input
  \item \code{idxb} vector \code{N} indices of samples statisfying the condition.  
}
You can easily derive the two datasets \code{Xb} and \code{Xnb} as: \code{Xb  = X[idxb, ]}, \code{Xnb = X[!idxb, ]}. 
If \code{Nboot > 1} it also contains
\itemize{ 
  \item \code{stat_lb} vector \code{(M)} lower bound of vector \code{(stat)} from bootstrapping
  \item \code{stat_ub} vector \code{(M)} upper bound of \code{stat} from bootstrapping vector
}
}
\description{
Computation function for Regional Sensitivity Analysis (with threshold). It splits the samples in a dataset \code{X} into two datasets (\code{Xb} and \code{Xnb}) depending on whether the associated sample in Y satisfies the condition:\eqn{Y(i, j) < threshold(j)}     for \eqn{j = 1,...,P}. Then assess the distance between the CDFs of \code{Xb} and \code{Xnb} by a suitable measure (maximum vertical distance, area between curves, etc.). Use the function \code{\link{RSA_plot_thres}} to visualize results.
}
\examples{
# See the demo
# demo("workflow_rsa_hymod")
}
\seealso{
\code{\link{RSA_plot_thres}} \code{\link{RSA_convergence_thres}}
}
