% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_plot_groups.R
\name{RSA_plot_groups}
\alias{RSA_plot_groups}
\title{Plotting function for Regional Sensitivity Analysis with grouping}
\usage{
RSA_plot_groups(X, idx, Yk, prnam = NULL)
}
\arguments{
\item{X}{matrix \code{(N, M)} set of input samples}

\item{idx}{vector \code{(N)} index of group to which input samples belong}

\item{Yk}{vector \code{(Ng + 1)} range of \code{Y} in each group}

\item{prnam}{vector \code{(M)} labels for the horizontal axis (default: \code{c("X1", "X2",...)})}
}
\description{
Plotting function for Regional Sensitivity Analysis with grouping. Plot \code{Ng} CDFs of the samples in \code{X} with different colours.
}
\examples{
# See the demo
# demo("workflow_rsa_hymod")
}
\seealso{
\code{\link{RSA_indices_groups}} \code{\link{RSA_plot_thres}}
}
