% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_plot_thres.R
\name{RSA_plot_thres}
\alias{RSA_plot_thres}
\title{Plotting function for Regional Sensitivity Analysis}
\usage{
RSA_plot_thres(X, idxb, prnam = NULL, threshold)
}
\arguments{
\item{X}{matrix \code{(N, M)} set of input samples}

\item{idxb}{vector \code{(N)} indices of samples statisfying the condition}

\item{prnam}{vector \code{(M)} labels for the horizontal axis (default: \code{c("X1", "X2",...)})}

\item{n_col}{scalar number of panels per row in the plot (default: \code{min(5,M)})}
}
\description{
This function plots Regional Sensitivity Analysis
}
\examples{
# See the demo
# demo("workflow_rsa_hymod")
}
\seealso{
\code{\link{RSA_indices_thres}} \code{\link{RSA_convergence_thres}}
}
