% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_g_function.R
\name{sobol_g_function}
\alias{sobol_g_function}
\title{Sobol' g-function}
\usage{
sobol_g_function(x, a)
}
\arguments{
\item{x}{vector (\code{M}) of inputs \eqn{x(1), x(2), ... x(M)}. \eqn{x(i)} ~ \eqn{Unif(0,1)} for all \eqn{i}}

\item{a}{vector (\code{M}) or scalar}
}
\value{
\code{y} scalar output. Two attributes \code{V} scalar output variance (exact value computed analytically), \code{Si_ex} vector (3), first-order sensitivity indices (exact value computed analytically)
}
\description{
Implements the Sobol' g-function, a standard benchmark function in
the Sensitivity Analysis literature (see for instance Sec. 3.6 in 
Saltelli et al. (2008)).
}
\examples{

a <- 9 # options for the (fixed) parameters
M <- 5 # options for the number of inputs
y <- sobol_g_function(runif(M), a)
Si_ex <- attributes(y)$Si_ex
}
\references{
Saltelli et al. (2008) Global Sensitivity Analysis, The Primer, Wiley.
}
