% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sample.R
\name{split_sample}
\alias{split_sample}
\title{Split sample}
\usage{
split_sample(Z, ngroup = 10)
}
\arguments{
\item{Z}{vector \code{(N, 1)}  sample of a model input  or output}

\item{ngroup}{number of groups considered (default: 10)}
}
\value{
List containing: 
\itemize{
  \item \code{idx} vector \code{N} respective group of the samples
  \item \code{Zk} vector \code{n_eff + 1} groups' edges (range of Z in each group)
  \item \code{Zc} vector \code{n_eff} groups' centers (mean value of Z in each group)
  \item \code{n_eff} scalar number of groups actually used to split the sample
}
}
\description{
Split a sample in n equiprobable groups based on the sample values (each groups contains approximately the same number of data points).
}
