% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbsa_convergence.R
\name{vbsa_convergence}
\alias{vbsa_convergence}
\title{Variance-based first-order and total effects indices}
\usage{
vbsa_convergence(Y, M, NN, Nboot = 0, alfa = 0.05)
}
\arguments{
\item{Y}{vector of \code{N * (M + 2)} model output samples  \code{Y = c(YA, YB, YC)}}

\item{M}{scalar number of inputs}

\item{NN}{vector \code{(R)}, sample sizes at which indices will be estimated (must be a vector of integer positive values not exceeding \code{N})}

\item{Nboot}{scalar, number of resamples used for boostrapping (default: 0)}

\item{alfa}{scalar, significance level for the confidence intervals estimated by bootstrapping (default: 0.05)}
}
\value{
List containing: 
\itemize{
  \item \code{Si} estimates of the main effects at different sampling size
  \item \code{STi}estimates of the total effects at different sampling size. 
}
If \code{Nboot > 1} it also contains
\itemize{
  \item \code{Si_sd} standard deviation of main effects at different sampling size
  \item \code{STi_sd} standard deviation of total effects at different sampling size
  \item \code{Si_lb} lower bound of main effects at different sampling size
  \item \code{STi_lb} lower bound of total effects at different sampling size
  \item \code{Si_ub} upper bound of main effects at different sampling size
  \item \code{STi_ub} upper bound of total effects at different sampling size. 
}
All output arguments are matrices of size \code{(R, M)}
}
\description{
This function computes the variance-based first-order indices
(or main effects) and total effects indices (Homma and Saltelli, 1996)
using an increasing number of output samples.
}
\examples{

fun_test  <- "ishigami_homma_function"
M <- 3 
distrfun <- "unif"
distrpar <- c(-pi, pi)
N <- 1000
sampstrat <- "lhs"
X <- AAT_sampling(sampstrat, M, distrfun, distrpar, 2 * N)
XABC <- vbsa_resampling(X)
YA <- model_evaluation(fun_test, XABC$XA)
YB <- model_evaluation(fun_test, XABC$XB)
YC <- model_evaluation(fun_test, XABC$XC)
Y <- c(YA, YB, YC)
NN <- seq(N/10, N, by = N / 10)
SiSTi <- vbsa_convergence(Y, M, NN)
}
\seealso{
\code{\link{vbsa_indices}} \code{\link{vbsa_resampling}}
}
