% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbsa_indices.R
\name{vbsa_indices}
\alias{vbsa_indices}
\title{Variance-based first-order indices and total effects indices}
\usage{
vbsa_indices(YA, YB, YC, Nboot = 0, alfa = 0.05)
}
\arguments{
\item{YA}{vector \code{(N)}, set of output samples}

\item{YB}{vector \code{(N)}, set of output samples (independent from \code{YA})}

\item{YC}{vector \code{(N*M)}, set of output samples from resampling (*)}

\item{Nboot}{scalar, number of resamples used for boostrapping (default: 0)}

\item{alfa}{scalar, significance level for the confidence intervals estimated by bootstrapping (default: 0.05)}
}
\value{
List containing: 
\itemize{
  \item \code{Si} estimates of the main effects at different sampling size
  \item \code{STi} estimates of the total effects at different sampling size
}
If \code{Nboot > 1} it also contains
\itemize{ 
  \item \code{Si_sd} standard deviation of main effects at different sampling size
  \item \code{STi_sd} standard deviation of total effects at different sampling size
  \item \code{Si_lb} lower bound of main effects at different sampling size
  \item \code{STi_lb} lower bound of total effects at different sampling size
  \item \code{Si_ub} upper bound of main effects at different sampling size
  \item \code{STi_ub} upper bound of total effects at different sampling size. 
}
All output arguments are matrices of size \code{(R, M)}
}
\description{
This function computes the variance-based first-order indices
(or main effects) and total effects indices (Homma and Saltelli, 1996).
The indices are approximated by the estimator suggested e.g. 
in Saltelli et al. (2008) and (2010).
}
\details{
NOTES:

(*) By default, here we use the estimators described by Saltelli et al.
    (2008) and Saltelli et al. (2010) (see comments in the code for
    specific references to the equations implemented here!).
    These are obtained from 3 sets of output samples (\code{YA}, \code{YB} and \code{YC}), 
    which are obtained by model evaluation against three input matrices 
    \code{XA}, \code{XB} and \code{XC} generated by the \code{\link{vbsa_resampling}} function: 
    - see example below about how to use \code{\link{vbsa_resampling}}
    - see help of \code{\link{vbsa_resampling}} to learn more about \code{XA}, \code{XB} and \code{XC}.

(**) If bootstrapping is used, \code{Si} and \code{STi} are the average of the
     respective \code{Nboot} estimates obtained at each bootstrap resampling.
}
\examples{

fun_test  <- "ishigami_homma_function"
M <- 3 
distrfun <- "unif"
distrpar  <- c(-pi, pi)
N <- 1000
sampstrat <- "lhs"
X <- AAT_sampling(sampstrat, M, distrfun, distrpar, 2 * N)
XABC <- vbsa_resampling(X)
YA <- model_evaluation(fun_test, XABC$XA)
YB <- model_evaluation(fun_test, XABC$XB)
YC <- model_evaluation(fun_test, XABC$XC)
SiSTi <- vbsa_indices(YA,YB,YC)
}
\references{
Homma, T. and A., Saltelli (1996). Importance measures in global 
sensitivity analysis of nonlinear models. 
Reliability Engineering & System Safety, 52(1), 1-17.

Saltelli et al. (2008), Global Sensitivity Analysis, The Primer, Wiley.

Saltelli et al. (2010), Variance based sensitivity analysis of model 
output. Design and estimator for the total sensitivity index, Computer 
Physics Communications, 181, 259-270.
}
\seealso{
\code{\link{vbsa_resampling}}
}
