% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbsa_resampling.R
\name{vbsa_resampling}
\alias{vbsa_resampling}
\title{Resampling strategy needed to build the approximators of the first-order and total order sensitivity indices}
\usage{
vbsa_resampling(X)
}
\arguments{
\item{X}{matrix \code{(NX, M)} of \code{NX} input samples}
}
\value{
List containing: 
\itemize{
  \item \code{XA} matrix \code{(N, M)} first \code{N = NX / 2} rows of \code{X}
  \item \code{XB} matrix \code{(N, M)} last \code{N = NX / 2} rows of \code{X}
  \item \code{XC} matrix \code{(N*M, M)} Block matrix of \code{M} recombinations of \code{XA} and \code{XB}, each block \code{XCi} is a \code{(N, M)} matrix whose columns are all taken from \code{XB} exception made for \code{i}-th, which is taken from \code{XA}.
}
}
\description{
This function implements the resampling strategy needed to build the
approximators of the first-order (main effects) and total order
sensitivity indices (e.g. Saltelli et al. 2008; 2010).
This function is meant to be used in combination with \code{\link{vbsa_indices}}.
}
\examples{

fun_test  <- "ishigami_homma_function"
M <- 3 
distrfun <- "unif"
distrpar  <- c(-pi, pi)
N <- 1000
sampstrat <- "lhs"
X <- AAT_sampling(sampstrat, M, distrfun, distrpar, 2 * N)
XABC <- vbsa_resampling(X)
}
\references{
Saltelli et al. (2008), Global Sensitivity Analysis, The Primer, Wiley.

Saltelli et al. (2010), Variance based sensitivity analysis of model 
output. Design and estimator for the total sensitivity index, Computer 
Physics Communications, 181, 259-270.
}
\seealso{
\code{\link{vbsa_indices}}
}
